refactor: consolidate module organization under `blocks` directories across kits
- Moved various submodules like ResNet, Swin, NanoChat, and Whisper into unified `blocks` directories for improved modularity and organization.
- Updated all references, `use` statements, and exports to reflect the new structure.
