+++ crates/bunsen/src/kits/bimm/resnet/blocks/downsample.rs
