+++ crates/bunsen/src/kits/bimm/swin/v2/blocks/mod.rs
@@ -1,0 +1,20 @@
+//! # `Swin` v2 blocks.
+
+pub(crate) mod swin_model;
+pub mod window_attention;
+
+mod block_sequence;
+mod patch_merge;
+mod swin_block;
+mod windowing;
+
+#[doc(inline)]
+pub use block_sequence::*;
+#[doc(inline)]
+pub use patch_merge::*;
+#[doc(inline)]
+pub use swin_block::*;
+#[doc(inline)]
+pub use swin_model::*;
+#[doc(inline)]
+pub use windowing::*;
