+++ crates/bunsen/src/kits/bimm/swin/v2/blocks/window_attention/mod.rs
@@ -1,10 +1,15 @@
 //! Window attention operations for Swin Transformer v2.
+
 mod attention;
 mod attention_mask;
 mod pos_bias;
 mod pos_grid;
 
+#[doc(inline)]
 pub use attention::*;
+#[doc(inline)]
 pub use attention_mask::*;
+#[doc(inline)]
 pub use pos_bias::*;
+#[doc(inline)]
 pub use pos_grid::*;
