+++ crates/bunsen/src/kits/bimm/swin/v2/blocks/window_attention/pos_grid.rs
