+++ crates/bunsen/src/kits/bimm/swin/v2/blocks/windowing.rs
