+++ crates/bunsen/src/kits/gpts/nanochat/blocks/mod.rs
@@ -1,0 +1,9 @@
+//! # `NanoChat` blocks.
+
+mod block;
+pub(crate) mod model;
+
+#[doc(inline)]
+pub use block::*;
+#[doc(inline)]
+pub use model::*;
