+++ crates/bunsen/src/kits/speech/whisper/mod.rs
@@ -6,3 +6,13 @@
 
 pub mod blocks;
 pub mod pretrained;
+
+#[doc(inline)]
+pub use blocks::{
+    Whisper,
+    WhisperApiConfig,
+    WhisperMeta,
+    WhisperStructuralConfig,
+};
+#[doc(inline)]
+pub use pretrained::PytorchWhisperScanner;
